requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorzero==2.0
Flask==2.2.2
gpiozero==1.6.2
heyoo==0.0.6
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psutil==5.9.4
python-dotenv==0.21.0
requests==2.28.1
urllib3==1.26.12
waitress==2.1.2
Werkzeug==2.2.2
